Add single-logout support
Authentik can be configured to trigger a backend logout using the URL
/lacon/oidc/backchannel-logout/

When a user logs in, we store the sid against their session key. When
SLO is triggered, we look up all of the sessions associated with that
sid and invalidate them.

When a user logs out from within nomnom we first do a regular logout
(just as a belt-and-braces in case the SLO mechanism doesn't work) and
then redirect to Authentik's session invalidation URL to trigger SLO on
all the other services.

Note that we don't use the SLO url from OIDC. That user interface is
hella confusing, asking if they want to log out of the individual
service or authentik or everything. Instead, we use Authentik's
invalidation URL, which essentially logs them out of Authentik directly,
which then triggers SLO for everything. The downside of this is that
after being logged out, they're redirected to concourse instead of
nomnom. However, after logging in again through concourse they can get
to nomnom fairly easily.